fixes #13483 - docker v2 - add api controller for manifests and update for tags
This commit contains some initial changes to support the docker v2 api
controllers.

This includes the introduction of a new controller (manifests) and modifications
to the existing docker tags controller.

With docker v2, the docker image APIs will no longer be needed; however,
removal of them and any associations will be done separately from this commit.