fixes #18035 - only add repos have access to
WORK IN PROGRESS... on issue 18035...

needs more testing... need to also make sure that the behavior makes sense...

e.g. if a user only has access to repo A through product A...
should they be able to see and add repo B from product B to a content view...
(note: ensure that repo b is already associated with another view...)

also... once these changes are in place...
does a search for content on repo B return any results?  (it shouldn't)